Refactored built function
The most significant change is the use of a dataOffset variable to track
where filters stop and data starts. To work, this requires that filters
always come before data. The upside is that it simplified much of the
code (besides, it'd probably be a confusing UI to have filters spread
all over the place).

We also now store days as integers (epoch) rather than strings. Changes
to other parts of the code will need to accommodate this micro
optimization.