Refactored showByMode
Split the creation of the table from the populating the table. This
appears to improve performance by 2-3x, but my version isn't equivalent
to the original. I need to finish the changes (re-apply the styles, fix
the date, ...) and see. There's still a noticeable lag with thousands of
items.

This change also accommodates the data structures introduced by the
previous commits.